// client/stocklinx app/src/redux/license/constant.ts
export enum licenseConst {
  FETCH_LICENSES_REQUEST = "FETCH_LICENSES_REQUEST",
  FETCH_LICENSES_SUCCESS = "FETCH_LICENSES_SUCCESS",
  FETCH_LICENSES_FAILURE = "FETCH_LICENSES_FAILURE",
  FETCH_LICENSE_REQUEST = "FETCH_LICENSE_REQUEST",
  FETCH_LICENSE_SUCCESS = "FETCH_LICENSE_SUCCESS",
  FETCH_LICENSE_FAILURE = "FETCH_LICENSE_FAILURE",
  CREATE_LICENSE_REQUEST = "CREATE_LICENSE_REQUEST",
  CREATE_LICENSE_SUCCESS = "CREATE_LICENSE_SUCCESS",
  CREATE_LICENSE_FAILURE = "CREATE_LICENSE_FAILURE",
  UPDATE_LICENSE_REQUEST = "UPDATE_LICENSE_REQUEST",
  UPDATE_LICENSE_SUCCESS = "UPDATE_LICENSE_SUCCESS",
  UPDATE_LICENSE_FAILURE = "UPDATE_LICENSE_FAILURE",
  REMOVE_LICENSE_REQUEST = "REMOVE_LICENSE_REQUEST",
  REMOVE_LICENSE_SUCCESS = "REMOVE_LICENSE_SUCCESS",
  REMOVE_LICENSE_FAILURE = "REMOVE_LICENSE_FAILURE",
  SET_LICENSE = "SET_LICENSE",
  CLEAR_LICENSE = "CLEAR_LICENSE",
  SET_LICENSES = "SET_LICENSES",
  CLEAR_LICENSES = "CLEAR_LICENSES",
}